Slurm does not accept floats
Former-commit-id: 96c3d58995710d6364dd72dace6be8b41139e749 [formerly 5e7577968ed4eef00ffc30bba383634fc77e844e]
Former-commit-id: ab041e7ac63d174ce9ad25e422e56bc30ebf96fd

+++ pymatgen/io/abinitio/qadapters.py
@@ -136,9 +136,9 @@ def timelimit_parser(s):
 
 def any2mb(s):
     if is_string(s):
-        return float(Memory.from_string(s).to("Mb"))
+        return int(Memory.from_string(s).to("Mb"))
     else:
-        return float(s)
+        return int(s)
 
 
 class MpiRunner(object):
